+++ docs-src/api-table/index.ts
@@ -1,2 +1,3 @@
 export { default as API } from './ApiTable.svelte';
 export * from './common-props';
+export * from './types';